Added a handful of new extensions and condtions
Added a working temporary action on TestButton to call look for item and then apply item to special currency slot item.

Note:
Element locations can hold trailing points and must be dealt with, so far the implementation is working and has an extra check for verifying we are on the correct item when calling AsyncTryClick() if the AsyncMoveMouse() fails.