Attempt #1 at solving broken crafting
Added Min-Max delay, this is for mouse movements
We move the mouse -> then we wait -> then we check and return if true in most cases otherwise we loop and try the move again.

Adjusted some login for key down/up (not sure if this will help ever but it shouldnt hurt)

Added async versions of get stash items just in case the need arises.